Sceptre configs and Cloudformation templates for deploying iAtlas APIs, Databases, and GitLab Runner to AWS. (#19)
This is designed to be deployed in two steps.
First, there is a manual deploy with Sceptre that will get the GitLab runner and it's dependencies deployed. Additionally, the HostedZone must be built and configured for the domain names and certs (see the README for more info).
Second, are deployments from the GitLab runner in CI/CD. The GitLab runner exists solely within the VPC and cannot be accessed publicly. It has access to the Database on the private subnet. The database cannot be accessed publicly. The API app itself is accessed publicly via a Load Balancer. This is the only public ingress.
The passwords for the database are created dynamically with the database and should never be exposed outside of the VPC.

This has been tested and is deployed in the GenUI AWS tenant: https://sage-api-staging.genui.dev/graphiql