Deploy with service user
The CI is setup to deploy using the collaborator's aws user/role.

IT removed the collaborator's access to the AWS account in PR https://github.com/Sage-Bionetworks-IT/organizations-infra/pull/407
Therefore it now fails to run deployments.

We don't want the CI to use the colloborator's user/role to deploy
resources to AWS.  Instead we want to use the service user/role to
do the deployment.  This PR switches to using the service user
to deploy to AWS.